speicher problem sollte behoben sein

+++ Programm/FWCP_MK3/FWCP_MK3/main.cpp
@@ -511,7 +511,7 @@ void maininterupthandler(monitor *mon){
 			Sekundenges += rtc.Minuten*60;
 			Sekundenges += rtc.Sekunden;
 			//GPS Daten und Zeit
-			f_printf(&logger,"%u\t%ld\t%ld\t%ld\t",(uint16_t)Sekundenges,(int64_t)(lon*1000000),(int64_t)(lat*100000),(int32_t)(gpsspeed*100));
+			f_printf(&logger,"%u\t%ld\t%ld\t%ld\t",(uint16_t)Sekundenges,(int32_t)(lon*1000000),(int32_t)(lat*100000),(int32_t)(gpsspeed*100));
 			//Tacho
 			f_printf(&logger,"%ld\t",(int32_t)(geschw*100));
 			//Barometer
